+++ packages/runtime/src/runtime/navigateApi.tsx
@@ -3,17 +3,7 @@ import type {
   ScreenContextDefinition,
 } from "@ensembleui/react-framework";
 import { isString } from "lodash-es";
-import type { NavigateFunction, Location } from "react-router-dom";
-
-export interface BrowserHistoryDefinition {
-  navigate: NavigateFunction | undefined;
-  location: Location | undefined;
-}
-
-export const browserHistory: BrowserHistoryDefinition = {
-  navigate: undefined,
-  location: undefined,
-};
+import { browserHistory } from "./history";
 
 export const navigateApi = (
   targetScreen: NavigateScreenAction,
